show auto reply depending on the MAP version and features
The version/feature is from an async call, to construct the notification
message, we also need an async call. So the solution now is to put
the feature result and pending notification in maps. This is an overly
complicated solusion for such a simple feature.

Ideally we want to put feature and version in MapCient.

Bug: 36987515

Test: manually verified

Change-Id: Ib7d1d0022df135c8434e21096699a1aa9b3a8944