Support receiving SMS messages through Telephony broadcasts.
Previously, this Messaging app was only receiving messages through
BluetoothMapClient. In order for this messaging app to function on the phone
as well, it must support receiving SMS messages through Telephony, and
writing the messages to the Telephony DB if its the default app.

This CL will enable us to use the same app on Embedded and Projected
platforms to test out Comm Center.

Fix: 129945932
Test: manual
Change-Id: I994b01842eda24dd9ab5e64ca045f0209c9f4fbe