Differentiate between message read on the phone, and on the car.
This is a fix for a bug that was occuring when the phone's messaging app was
open when a message was received, causing CarMessenger to believe message
was read, which resulted in empty notifications (since only unread messages
are included in the notification).

We should still track whether it's read on the phone, especially when loading
existing messages, future comm center support. However, changing our design
such that, message notifications only contain messages that haven't been
read in the CAR only.

Also minor fix in MessengerDelegateTest to fix flakiness.

Change-Id: I4a82ea4c735e9ac01fa0bc7d3986404b346cea41
Fix: 138737376
Test: unit test, verified on hawk.