Disconnect from MAP on onDestroy
This CL fixes a couple of bugs:
1. When Messenger gets onDestroy we should call closeProfileProxy as that
is what is causing ServiceConnection leaks
2. onServiceDisconnected is called when the connection is lost, not necessarily
when the connection is closed so instead of calling BMC.close in this situation,
we should only set BMC to null and not make any BMC API calls until
onServiceConnected
3. Because of 2, trying to reconnect by calling getProfileProxy won't speed
up getting the BMC, so we should stop doing that.
4. Sal mentioned most BT lifecycle logic relies on closeProfileProxy, not
BMC.close, so I stopped calling that API
5. For readability's sake, I renamed the cleanup() methods to specifically
onDestroy so readers can understand when this method is invoked.

Change-Id: I7727df11664a67b809a2361289afbc4623877631
Fix: 142077620
Test: manual, force onDestroy