Exclude dismissed and read messages from the notification
When a user dismisses a message notification, OR when a message
notification is successfully read aloud, the messages in that notification
should not appear when/if that notification gets updated with newer
messages.

This change combines the two states (dismiss and read) since they both
have the same logical consequence.

Change-Id: I1518a86606f4035e6f721844e4705442482a3307
Fix: 146362547
Test: manual