Fix group message title when switching to RTL languages
Generated titles should not be reused after changing the locale.
Because:
1 - Change of locale means none of the separators used are valid
anymore. For example a string like "name, name2, name3" cannot be used
in a language that does not use "," as a separator.
2 - All the text direction markers should be reassessed when the locale
changes. Please note that TextDirectionHeuristics.LOCALE translates
to either RLM or LRM based on the current locale. The generated string
should not be used in another locale with a different text direction.

Bug: 159221409
Bug: 161947401
Test: Manual
Change-Id: I725f23a306ff97889d08879c44a38c94f128d326